Rename GenerateMailTemplatesService to MailThemeGenerator
